+++ lib/pescarte/modulo_pesquisa/repository.ex
@@ -129,6 +129,7 @@ defmodule Pescarte.ModuloPesquisa.Repository do
   @impl true
   def fetch_relatorio_pesquisa_by_id(id_publico) do
     Repo.replica().get_by(RelatorioPesquisa, id_publico: id_publico)
+    |> Repo.replica().preload(pesquisador: [:linha_pesquisa, usuario: [contato: [:endereco]], orientador: [:usuario]])
   end
 
   @impl true
