+++ lib/pescarte_web/templates/relatorio_html/relatorio_anual.html.heex
@@ -30,6 +30,14 @@
           margin-inline: auto;
       }
 
+      .capa__cidade_data {
+        width: 100%;
+        display: flex; 
+        justify-content: center;  
+        position: absolute; 
+        bottom: 0; 
+      }
+
       .sumario {
           width: 18cm;
           height: 25.2cm;
@@ -86,9 +94,10 @@
         <b><%= build_usuario_name(@pesquisador.usuario) %></b>
       </p>
       <p><b>Pesquisador (a)</b></p>
-      <p style="position: absolute; bottom: 0; left: 6.25cm">
+      <p class="capa__cidade_data">
         <b>
-          <%= ~c"#{@pesquisador.usuario.contato.endereco.cidade}/#{@pesquisador.usuario.contato.endereco.estado}" %><br /><%= get_literal_mes(@data_fim) %>/<%= @data_entrega.year %>
+          <%= ~c"#{@pesquisador.usuario.contato.endereco.cidade}/#{@pesquisador.usuario.contato.endereco.estado}" %>
+          <%= get_literal_mes(@data_fim) %>/<%= @data_entrega.year %>
         </b>
       </p>
     </section>
